refactor tests/util to tests/test_utils
