+++ tests/column_map_expectations/test_column_map_expectations.py
@@ -14,7 +14,7 @@ import json
 import glob
 import warnings
 
-from tests.util import get_dataset, candidate_test_is_on_temporary_notimplemented_list, evaluate_json_test
+from tests.test_utils import get_dataset, candidate_test_is_on_temporary_notimplemented_list, evaluate_json_test
 
 contexts = ['PandasDataSet', 'SqlAlchemyDataSet']
